requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
comm==0.1.4
cryptography==39.0.1
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.1
-e git+https://github.com/R-I-S-Khan/FedScaleImp.git@f9e6863001aa9299e6bce759bd96a9ad5036be53#egg=fedscale
filelock==3.12.2
flatbuffers==23.5.26
flit-core==3.6.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
gdown==4.7.1
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.40.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
HeapDict==1.0.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==4.11.3
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==4.12.0
jupyter-server==1.24.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.6
librosa==0.7.2
llvmlite==0.32.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.8
notebook==6.5.6
notebook-shim==0.2.3
numba==0.49.1
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==23.2
pandas==1.1.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.11.1
pycparser==2.21
pygments==2.17.2
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.12.0
pytorch-ranger==0.1.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
redis==3.5.3
redis-py-cluster==2.1.3
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.0
resampy==0.4.2
rsa==4.7.2
safetensors==0.4.1
scikit-learn==1.0.2
scipy==1.4.1
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
sox==1.3.7
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
torch==1.13.1
torch-baidu-ctc==0.3.0
torch-optimizer==0.3.0
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
typing-extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.2
widgetsnbextension==4.0.9
wrapt==1.16.0
yarl==1.8.1
zipp==3.11.0
